refactor(frontend): Se ha modularizado la función mis vuelos.
