+++ frontend/src/utils/store.js
@@ -7,3 +7,4 @@ export let mensaje_exito = writable(null);
 export let mensaje_error = writable(null);
 
 
+
